+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 ## Unreleased
 - [Feature] Added new Storybook examples detailing Flexbox helper classes ([#1367](https://github.com/optimizely/oui/pull/1367))
 - [Feature] Add ability to sort table columns to **Table.TH** ([#1365](https://github.com/optimizely/oui/pull/1365))
+- [Feature] Add `textAlign` to **Table.TH** & **Table.TD** ([#1370](https://github.com/optimizely/oui/pull/1370))
 
 ## 46.8.0 - 2020-06-30
 - [Feature] Add ability to supply custom links to **ManagerSideNav** and style with custom classes ([#1362](https://github.com/optimizely/oui/pull/1362))